fix: change evm provider used by evmTools
This refactors the evm tools to resolve (hopefully) a problem that occurs
when a user has multiple ethereum providers. Instead of using
`window.ethereum` directly, get the provider from the wallet store.

This guarantees that even if a different wallet wins the race to set
`window.ethereum`, we still use the correct provider. In environments
where multiple wallets are connected, `window.ethereum` may no work as
we previously expected.