refactor & reimplement evm tools and chain registry (#134)
* entirely reimplement evm tool calls

* replace previous tool call implementation

* refactor and test improvement wip

* refactor: extract evmTools into multiple files

* fix: improve agent ability to format ETH correctly

* feat: improve provided connected wallet context

* debug: add a debug button that outputs chat to console

* fix: chain lookup by hex chain id

* update chain name tool call description to omit "Kava EVM"

* fix: change evm provider used by evmTools

This refactors the evm tools to resolve (hopefully) a problem that occurs
when a user has multiple ethereum providers. Instead of using
`window.ethereum` directly, get the provider from the wallet store.

This guarantees that even if a different wallet wins the race to set
`window.ethereum`, we still use the correct provider. In environments
where multiple wallets are connected, `window.ethereum` may no work as
we previously expected.